Added a pricing service to do operatiosn on pricing data
