+++ FeatureRecognitionAPI/FeatureRecognitionAPI/Services/IPricingDataService.cs
@@ -1,0 +1,10 @@
+using FeatureRecognitionAPI.Models.Dtos;
+using FeatureRecognitionAPI.Models.Pricing;
+
+namespace FeatureRecognitionAPI.Services;
+
+public interface IPricingDataService
+{
+    PunchPriceReturn GetPunchPrices();
+    List<FeaturePrice> GetFeaturePrices();
+}
